Prefer Set<K> over Iterable<K> for callback keys parameter
An Iterable<K> should be preferred as a parameter type for methods that
users may call, such as getAll(keys). This way it can accept the widest
range of collection types and the cache will remove duplicate.

The callback methods to user code, however, should use the most
specific type, such as CacheLoader.loadAll(keys). This way the
implementor can better use the parameter rather than copying into a
more usable collection type. As only the cache invokes the callbacks,
it can pass the Set through without any boilerplate on either side.

This api mistake exists in Guava, which Caffeine inherited. The impact
is minor and only for bulk loader implementations. In use cases like
getAll(keys, func), likely only a recompilation is needed if the types
are inferred.